server/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-cors-headers==3.10.0
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
numpy==1.21.3
pandas==1.3.4
psycopg2==2.9.1
PyJWT==1.7.1
python-dateutil==2.8.2
python-http-client==3.3.3
pytz==2021.3
sendgrid==6.8.3
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==1.1.1
